… .github/workflows/comprehensive-tests.yml
… .github/workflows/docker-tests.yml
… .github/workflows/hybrid-runner-strategy.yml
… .github/workflows/kubernetes-tests.yml
… .github/workflows/selective-tests.yml
… .github/workflows/self-hosted-tests.yml
… docs/SELF_HOSTED_RUNNERS.md
… test/regression/__snapshots__/health-report.json
… test/regression/__snapshots__/hybrid-strategy-latest.json
… test/security/prompt-injection.test.js