## Summary
### Part 1: Test Fixes and Performance Improvements ✅
- Fixed failing security tests and dramatically improved test performance
- Integration tests now run 375x faster (60s → 160ms)
- All security tests passing successfully

### Part 2: Self-Hosted Runner Support 🚀
- Added comprehensive self-hosted runner support for all workflows
- Created hybrid runner strategy with automatic fallback
- Complete documentation and setup guide

## Changes

### Test Fixes (Part 1)
- **Fixed prompt injection detection** in `hybrid-strategy.test.js`
- **Enhanced role manipulation detection** in `prompt-injection.test.js`
- **Optimized integration test performance**: 60s → 160ms (375x faster)

### Self-Hosted Runners (Part 2)

#### New Workflows
- `self-hosted-tests.yml`: Dedicated self-hosted runner workflow
- `hybrid-runner-strategy.yml`: Intelligent runner detection and fallback

#### Updated Workflows
- `comprehensive-tests.yml`: Hybrid runner support
- `docker-tests.yml`: Prefers self-hosted for Docker
- `kubernetes-tests.yml`: Optimized for K8s tools

#### Documentation
- `docs/SELF_HOSTED_RUNNERS.md`: Complete setup guide
- Installation scripts
- Troubleshooting
- Cost analysis (90% savings)

## Benefits
- **Performance**: 375x faster execution
- **Cost**: ~$20/month vs $200 GitHub runners
- **Flexibility**: Automatic fallback
- **Caching**: Persistent dependencies

## Test Status
✅ All tests passing with optimized performance

🤖 Generated with [Claude Code](https://claude.ai/code)