@rafeekpro 2025-09-14
## ✅ Test Fixes Complete

All test failures have been resolved locally:
- ✅ **Regression tests**: Fixed file path validation issues  
- ✅ **Installation tests**: Fixed AWK error and input handling
- ✅ **Security tests**: All passing
- ✅ **Performance tests**: All passing

### ⚠️ GitHub Actions Queue Issue
The CI/CD checks are currently stuck in pending state (>5 minutes) waiting for GitHub-hosted runners. This appears to be either:
1. GitHub Actions monthly quota limit reached
2. GitHub infrastructure temporary issue

### Local Test Results
```
npm run test:all
✅ All tests passing locally
```

The code changes are ready for merge. All fixes have been verified locally.